+++ data/duke.txt
@@ -1,3 +1,5 @@
 D | 0 | Test2 | 2022-10-15
 E | 0 | Test3 | 2023-12-30
-T | 0 | test4
+
+E | 0 | Test4 | 2023-12-12
+D | 0 | Test5 | 1998-12-01
