+++ src/main/java/duke/Duke.java
@@ -1,0 +1,32 @@
+package duke;
+
+import duke.command.Command;
+
+public class Duke {
+    static String path2 = "data/duke.txt";
+
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    public Duke(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        tasks = new TaskList(storage.load());
+    }
+
+    public void run() {
+        ui.showWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            String fullCommand = ui.readCommand();
+            Command c = Parser.parseCommand(fullCommand);
+            c.execute(tasks, ui, storage);
+            isExit = c.isExit();
+        }
+    }
+
+    public static void main(String[] args) {
+        new Duke(path2).run();
+    }
+}
