+++ src/main/java/duke/Parser.java
@@ -1,4 +1,6 @@
-import java.io.IOException;
+package duke;
+
+import duke.command.*;
 
 public class Parser {
