+++ src/main/java/duke/Storage.java
@@ -1,9 +1,15 @@
+package duke;
+
+import duke.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
+
 import java.util.ListIterator;
 import java.util.Scanner;
+
 
 public class Storage {
     private File homeFile = new File("data");
