+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,8 @@
-import java.io.IOException;
+package duke;
+
+import duke.exception.EmptyDescException;
+import duke.task.*;
+
 import java.util.ArrayList;
 import java.util.ListIterator;
 import java.util.Scanner;
