+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,20 +1,35 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+import duke.task.Deadline;
+
+import duke.exception.EmptyDescException;
+
 import java.io.IOException;
 
-public class EventCommand extends Command {
+public class DeadlineCommand extends Command {
     private String desc;
-    private String at;
+    private String by;
 
-    public EventCommand(String desc) {
-        this.desc = desc.replace("event", "");
+    public DeadlineCommand(String desc) {
+        desc = desc.replace("deadline", "");
         int slash = desc.lastIndexOf("/");
-        this.at = desc.substring(slash + 4);
-        this.desc = this.desc.substring(0, slash - 1);
+        this.by = desc.substring(slash + 4);
+        this.desc = desc.substring(0, slash - 1);
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            Event task = new Event(desc, at);
+            Deadline task = new Deadline(desc, by);
             tasks.addTask(task);
             storage.appendToFile(task.toSave() + System.lineSeparator() + "");
             ui.printMessage("Got it. I've added this task:\n"
