+++ src/main/java/duke/command/MarkCommand.java
@@ -1,10 +1,16 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
 
     private int index;
 
-    public UnmarkCommand(String desc) {
+    public MarkCommand(String desc) {
         String numberOnly = desc.replaceAll("[^0-9]", "");
         index = Integer.parseInt(numberOnly);
     }
@@ -12,9 +18,9 @@ public class UnmarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            tasks.unmarkTask(index);
+            tasks.markTask(index);
             storage.rebuildFile(tasks.getIterator());
-            ui.printMessage("Alright, this task is marked as not done yet:\n"
+            ui.printMessage("Alright, this task is marked as done:\n"
                     + tasks.getTask(index).toString());
         } catch (IOException ioe) {
             System.out.println("Something went wrong: " + ioe.getMessage());
