+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,14 +1,18 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.EmptyDescException;
+import duke.task.ToDo;
+
 import java.io.IOException;
 
-public class DeadlineCommand extends Command {
+public class ToDoCommand extends Command {
     private String desc;
-    private String by;
 
-    public DeadlineCommand(String desc) {
-        this.desc = desc.replace("deadline", "");
-        int slash = desc.lastIndexOf("/");
-        this.by = desc.substring(slash + 4);
-        this.desc = this.desc.substring(0, slash - 1);
+    public ToDoCommand(String desc) {
+        this.desc = desc.replace("todo", "");
     }
 
     @Override
@@ -19,9 +23,9 @@ public class DeadlineCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            Deadline task = new Deadline(desc, by);
+            ToDo task = new ToDo(desc);
             tasks.addTask(task);
-            storage.appendToFile(task.toSave() + System.lineSeparator() + "");
+            storage.appendToFile(task.toSave() + System.lineSeparator());
             ui.printMessage("Got it. I've added this task:\n"
                     + task + "\n"
                     + "Now you have " + tasks.listSize() + " tasks in the list.");
