+++ src/main/java/duke/exception/UnknownInputException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class UnknownInputException extends Exception {
     public UnknownInputException(String message) {
         super(message);
