+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.exception.EmptyDescException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
