+++ src/main/java/duke/task/ToDo.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.exception.EmptyDescException;
+
 public class ToDo extends Task {
     public ToDo(String desc) throws EmptyDescException {
         super(desc);
